feat : 학생 내 정보 조회 API 개발

+++ src/main/java/com/monari/monariback/enrollment/dto/request/EnrollmentCreateRequest.java
@@ -2,7 +2,7 @@ package com.monari.monariback.enrollment.dto.request;
 
 public record EnrollmentCreateRequest(
     // TODO : Jwt 적용
-    Long studentId,
+    Integer studentId,
     Integer lessonId
 ) {
